SystemInternal and WorldControl into own headers
Reduces the size of simulationrunner header a bit
